Functionality and design updates
Fixed scaling of the nav menu for smaller screens, added a hamburger menu.
Fixed issue with project images being cropped.
Reworked the display of featured projects into two rows of 3 instead of 1 row of 5.
Improved visibility of the Linkedin link in the contact section.
Added feature so that the nav menu displays which section the user is currently on.