Update to menu and image handling
Changed it so that the two different images for each project alternate on hover for desktop and on click for mobile. Also updated handling of going to the github repo link worked on mobile so it won't interfere for the onclick of changing image.
This update broke the grid layout and text box coherence, will fix this after.